API support for news items, refactoring events API controller
